🚚 Rename: ResponseCode.java 파일 이동 - 'global/exception' -> 'global/common'
